Created custom validators
Generated a class-validator for ensuring password and confirm-password registration fields match.
Created a static method within the RegisterValidators class to ascertain the match between the password and confirm-password and imported it in the registration component class.
Created a factory function to avoid hard-coding the password and confirm_password attributes into the method.
Updated the error handling process to include 'Passwords do not match' message beneath the confirm password field.
Created a separate class file to hold an asynchronous validator that verifies if an email has already been taken.
Updated registration template to feature the new errors.