Routing Commit 1
Created components to register the route for the home page and about page, added navigation to the app (updated navigation template file).

Added styling to the links based on the current path.

Generated a routing module for the video feature. Created several files related to the creation and management of videos.

Generated a component for rendering the page content.

Transferred the logout function from the nav class component to the auth service class.

Filtered the router events to feature only the navigation end object.

Refined the logout behaviour by adding data to a route which aids the redirection of the user if a certain component requires authentication to access.

Generated a new component for the upload functionality of the app and connected it via the router.

Generated a component to house the content for the video page and implemented a single route for multiple clips.

Synchronized the route to feature the correct title of the specific clip whenever the user navigates to a different clip.

Added query parameters to store the sorting order of videos on the manage page.

Fixed value persistence of recent and oldest uploads in the manage tab after page refresh.

Implemented a wildcard route for pages that cannot be found.