Updated upload functionality 1
Prevented the default browser behaviour when dragging files into the app by designing a custom directive.

Fixed the hover effect of the drop box (previously it would not change colour when a file was hovered over it). Added listeners on the drop box element of the upload template.

Implemented file check before upload and hid the 'select a thumbnail' section.