Updated upload functionality 2
Used reactive forms to make the upload form functional. Registered the reactive form with the upload component.

Connected the file name to the title field and implemented a method that removes file extension (.mp4) from the title.

Installed uuid and used it to create unique ids for the corresponding videos that are being uploaded.

Added an alert component to display the status of the upload.

Created an upload progress observable and used it to show the percentage of the upload progress. Created properties representing the current status of the upload. Bound said properties to the template. Changed the properties in the appropriate functions.